Final commit

+++ src/main/kotlin/com/jvcs/spring_boot_crash_course/controllers/NoteController.kt
@@ -3,7 +3,10 @@ package com.jvcs.spring_boot_crash_course.controllers
 import com.jvcs.spring_boot_crash_course.controllers.NoteController.NoteResponse
 import com.jvcs.spring_boot_crash_course.database.model.Note
 import com.jvcs.spring_boot_crash_course.database.repository.NoteRepository
+import jakarta.validation.Valid
+import jakarta.validation.constraints.NotBlank
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -28,6 +31,7 @@ class NoteController(
 
     data class NoteRequest(
         val id: String?,
+        @field:NotBlank(message = "Title can't be blank")
         val title: String,
         val content: String,
         val color: Long
@@ -44,8 +48,9 @@ class NoteController(
     @OptIn(ExperimentalTime::class)
     @PostMapping
     fun save(
-        @RequestBody body: NoteRequest
+        @Valid @RequestBody body: NoteRequest
     ): NoteResponse {
+        val ownerId = SecurityContextHolder.getContext().authentication?.let { it.principal as String }
         val note =  noteRepository.save(
             Note(
                 id = body.id?.let { ObjectId(it)} ?: ObjectId.get(), // if id is null post a new note, if it is not null create a new note
@@ -53,7 +58,7 @@ class NoteController(
                 content = body.content,
                 color = body.color,
                 createdAt = Clock.System.now(),
-                ownerId = ObjectId()
+                ownerId = ObjectId(ownerId)
             )
         )
 
@@ -61,9 +66,8 @@ class NoteController(
     }
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String
-    ): List<NoteResponse> {
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication?.let { it.principal as String }
         return noteRepository.findByOwnerId(ObjectId(ownerId)).map {
             it.toNoteReponse()
         }
@@ -71,7 +75,13 @@ class NoteController(
 
     @DeleteMapping(path = ["/{id}"])
     fun deleteById(@PathVariable id: String) {
-        noteRepository.deleteById(ObjectId(id))
+        val note = noteRepository.findById(ObjectId(id)).orElseThrow {
+            throw IllegalArgumentException("Note not found.")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication?.let { it.principal as String }
+        if (note.ownerId.toHexString() == ownerId ) {
+            noteRepository.deleteById(ObjectId(id))
+        }
     }
 
 }
